Refactor config properties.
