+++ backend/src/main/java/ca/metricalsky/winston/config/api/VideosApiConfig.java
@@ -1,4 +1,4 @@
-package ca.metricalsky.winston.config;
+package ca.metricalsky.winston.config.api;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
@@ -8,7 +8,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
-@ConfigurationProperties(prefix = "api.videos")
+@ConfigurationProperties(prefix = "winston.api.videos")
 @Value
 public class VideosApiConfig {
