+++ backend/src/main/resources/application.properties
@@ -1,9 +1,9 @@
-api.videos.default-page-size=10
 management.endpoints.web.exposure.include=health,info
 spring.application.name=winston
 spring.flyway.clean-disabled=true
 spring.jackson.default-property-inclusion=non_null
 spring.jpa.hibernate.ddl-auto=validate
 spring.main.banner-mode=off
+winston.api.videos.default-page-size=10
 youtube.apiKey=${YOUTUBE_API_KEY}
 youtube.quota.daily=10000
