+++ backend/src/test/resources/application.properties
@@ -1,9 +1,9 @@
-api.videos.default-page-size=10
 spring.datasource.url=jdbc:postgresql://localhost:5432/yt-comments-test
 spring.flyway.clean-disabled=false
 spring.flyway.validate-migration-naming=true
 spring.jpa.hibernate.ddl-auto=validate
 spring.test.database.replace=none
+winston.api.videos.default-page-size=10
 youtube.apiKey=YOUTUBE_API_KEY
 youtube.mock.rootUrl=${wiremock.server.baseUrl}
 youtube.quota.daily=10000
