+++ frontend/package.json
@@ -24,7 +24,7 @@
     "react-bootstrap-icons": "^1.11.6",
     "react-dom": "^19.1.1",
     "react-redux": "^9.2.0",
-    "react-router": "^7.9.1",
+    "react-router": "^7.9.3",
     "react-sse-hooks": "^1.0.6",
     "uuid": "^13.0.0"
   },
@@ -39,11 +39,11 @@
     "@types/bootstrap": "^5.2.10",
     "@types/lodash": "^4.17.20",
     "@types/luxon": "^3.7.1",
-    "@types/react": "^19.1.13",
+    "@types/react": "^19.1.14",
     "@types/react-dom": "^19.1.9",
-    "@vitejs/plugin-react": "^5.0.3",
+    "@vitejs/plugin-react": "^5.0.4",
     "@vitest/coverage-v8": "^3.2.4",
-    "@vitest/eslint-plugin": "^1.3.12",
+    "@vitest/eslint-plugin": "^1.3.13",
     "eslint": "^9.36.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-import-resolver-vite": "^2.1.0",
